feat: SurveyPage 컴포넌트로 분리 (#95)
