+++ src/components/ui/carousel.tsx
@@ -230,6 +230,7 @@ CarouselNext.displayName = 'CarouselNext';
 
 export {
   type CarouselApi,
+  useCarousel,
   Carousel,
   CarouselContent,
   CarouselItem,
